+++ templates/default/fulldoc/html/setup.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+def generate_assets
+  super
+  asset('css/kward.css', file('css/kward.css', true))
+  asset('js/kward.js', file('js/kward.js', true))
+  asset('images/kward_logo.png', file('images/kward_logo.png', true))
+end