+++ templates/default/layout/html/setup.rb
@@ -1,0 +1,110 @@
+# frozen_string_literal: true
+
+module KwardDocsNavigation
+  GUIDE_GROUPS = [
+    [
+      "Start here",
+      [
+        ["Getting started", "file.getting-started.html"],
+        ["Usage", "file.usage.html"],
+        ["Configuration", "file.configuration.html"],
+        ["Authentication", "file.authentication.html"],
+        ["Troubleshooting", "file.troubleshooting.html"]
+      ]
+    ],
+    [
+      "Feature guides",
+      [
+        ["Memory", "file.memory.html"],
+        ["Personas", "file.personas.html"],
+        ["Extensibility", "file.extensibility.html"],
+        ["Plugins", "file.plugins.html"],
+        ["Web search", "file.web-search.html"],
+        ["Code search", "file.code-search.html"]
+      ]
+    ],
+    [
+      "Advanced/reference",
+      [
+        ["RPC protocol", "file.rpc.html"],
+        ["Releasing", "file.releasing.html"]
+      ]
+    ]
+  ].freeze
+
+  GUIDE_OVERVIEW = "file.README.html"
+  GUIDE_LINKS = ([GUIDE_OVERVIEW] + GUIDE_GROUPS.flat_map { |_title, items| items.map(&:last) }).freeze
+  GUIDE_FILE_LINKS = {
+    "doc/getting-started.md" => "file.getting-started.html",
+    "doc/usage.md" => "file.usage.html",
+    "doc/configuration.md" => "file.configuration.html",
+    "doc/authentication.md" => "file.authentication.html",
+    "doc/troubleshooting.md" => "file.troubleshooting.html",
+    "doc/memory.md" => "file.memory.html",
+    "doc/personas.md" => "file.personas.html",
+    "doc/extensibility.md" => "file.extensibility.html",
+    "doc/plugins.md" => "file.plugins.html",
+    "doc/web-search.md" => "file.web-search.html",
+    "doc/code-search.md" => "file.code-search.html",
+    "doc/rpc.md" => "file.rpc.html",
+    "doc/releasing.md" => "file.releasing.html"
+  }.freeze
+
+  def guide_groups
+    GUIDE_GROUPS
+  end
+
+  def guide_overview
+    GUIDE_OVERVIEW
+  end
+
+  def current_docs_path
+    serializer = options.serializer
+    return "index.html" unless serializer
+
+    File.basename(serializer.serialized_path(object))
+  end
+
+  def readme_file?
+    defined?(@file) && @file&.name == "README"
+  end
+
+  def guide_file?
+    defined?(@file) && GUIDE_FILE_LINKS.key?(@file&.filename.to_s)
+  end
+
+  def home_page?
+    options.index && readme_file?
+  end
+
+  def guide_page?
+    (readme_file? && !options.index) || guide_file? || GUIDE_LINKS.include?(current_docs_path)
+  end
+
+  def api_page?
+    !home_page? && !guide_page?
+  end
+
+  def diskfile
+    @file.attributes[:markup] ||= markup_for_file('', @file.filename)
+    data = rewrite_guide_links(htmlify(@file.contents, @file.attributes[:markup]))
+    "<div id='filecontents'>" + data + "</div>"
+  end
+
+  def rewrite_guide_links(html)
+    GUIDE_FILE_LINKS.each do |source, target|
+      html = html.gsub(%(href="#{source}"), %(href="#{target}"))
+    end
+    html
+  end
+end
+
+include KwardDocsNavigation
+
+def javascripts
+  super + %w(js/kward.js)
+end
+
+def stylesheets
+  super + %w(css/kward.css)
+end